requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
mysql-connector-python==8.2.0
